condaenv.r8ogg3eu.requirements.txt
openai
python-dotenv
pydantic
fastapi
langtrace-python-sdk
requests
ipykernel